Fix storage comparison typings and style consistency

+++ storage_comparison/clickhouse_ch/clickhouse_storage.py
@@ -14,7 +14,7 @@ class ClickhouseStorage(AbstractStorage):
         self.client = client
         self.generator = generator
 
-    def _execute_query(self, query: str) -> int:
+    def _execute_query(self, query: str) -> float:
         start = time()
         self.client.execute(query)
         end = time()
@@ -29,14 +29,14 @@ class ClickhouseStorage(AbstractStorage):
         query = f'DROP TABLE {name};'
         self._execute_query(query)
 
-    def add(self, query: str, table: str, total: int, batch: int) -> int:
+    def add(self, query: str, table: str, total: int, batch: int) -> float:
         start = time()
         for q in self.generator.add_data(total, batch, query, table):
             self.client.execute(q)
         end = time()
         return end - start
 
-    def read(self, query: str, table: str) -> int:
+    def read(self, query: str, table: str) -> float:
         query = query % table
         result = self._execute_query(query)
         return result
